add test_full_ar6
this test provides a simple test case which should cover the species that should be harmonized for AR6.

Notes and issues raised are the sum of what was raised in the previous two commits:

Note: no overrides provided.
Note: test_pipeline_progress.xlsx values that are there have thusfar not been checked for quality (just output of default hamonization run).

Issues raised here:

- the binary 'global_harmonization_only' switch chooses for either taking in sector-only, or aggregate-only data. Hence, not all variables will be harmonized here.

- in either binary setting, no data is given for the region&sector data; hence the ?? values in the test_pipeline_progress.xlsx file. [note that the harmonization (for the right binary switch mode) does supply harmonized data for CH4, but not for Sulfur]